carimbo-service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
opencv-python-headless>=4.10,<5.0
Pillow>=10.4,<12.0
pdf2image>=1.17,<2.0
numpy>=1.26,<3.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
python-multipart>=0.0.9,<1.0
pytest>=8.2,<9.0
